Fixed /archive, added alias /world
Theory: Bukkit.getWorld() returns null when used on a scaffold world in a closed state. I removed this line of code as well as world.save() to fix the command. Note: I'm pretty confident world.save() was redundant, so this should not mess up archived worlds.

Also added /world alias to the /open command